+++ RELEASE-NOTES.md
@@ -4,7 +4,7 @@
 ## Deprecations
 
 ## New additions
-
+ * Add external access integration to snow object commands
 ## Fixes and improvements
 
 
@@ -21,6 +21,7 @@
   * Passing files and directories as arguments syncs these only: `snow app deploy some-file some-dir`.
   * `--recursive` syncs all files and subdirectories recursively.
   * `--prune` deletes specified files from the stage if they don't exist locally.
+* Added the `--cascade` option to `snow app teardown` command that drops all application objects owned by the application.
 
 ## Fixes and improvements
 * More human-friendly errors in case of corrupted `config.toml` file.